+++ app/src/main/java/com/xuexiang/xuidemo/base/webview/AgentWebFragment.java
@@ -700,10 +700,6 @@ public class AgentWebFragment extends Fragment implements FragmentKeyDown {
      * @return
      */
     protected MiddlewareWebClientBase getMiddlewareWebClient() {
-        if (LollipopFixedWebView.isLollipopWebViewBug()) {
-            return null;
-        }
-
         return new MiddlewareWebViewClient() {
             /**
              *
